config abstraction
