+++ gonk-hotkeys/src/main.rs
@@ -1,4 +1,4 @@
-use gonk_core::HotkeyConfig;
+use gonk_core::{Config, GlobalHotkey, HOTKEY_CONFIG};
 use gonk_server::Client;
 use win_hotkey::Listener;
 
@@ -13,32 +13,29 @@ enum Event {
 }
 
 fn main() {
-    let config = HotkeyConfig::new();
+    let config: Config<GlobalHotkey> = Config::new(HOTKEY_CONFIG.as_path());
     let mut hk = Listener::<Event>::new();
     let mut client = Client::new();
+    let data = config.data;
 
     hk.register_hotkey(
-        config.volume_up.modifiers(),
-        config.volume_up.key(),
+        data.volume_up.modifiers(),
+        data.volume_up.key(),
         Event::VolUp,
     );
     hk.register_hotkey(
-        config.volume_down.modifiers(),
-        config.volume_down.key(),
+        data.volume_down.modifiers(),
+        data.volume_down.key(),
         Event::VolDown,
     );
+    hk.register_hotkey(data.previous.modifiers(), data.previous.key(), Event::Prev);
+    hk.register_hotkey(data.next.modifiers(), data.next.key(), Event::Next);
     hk.register_hotkey(
-        config.previous.modifiers(),
-        config.previous.key(),
-        Event::Prev,
-    );
-    hk.register_hotkey(config.next.modifiers(), config.next.key(), Event::Next);
-    hk.register_hotkey(
-        config.play_pause.modifiers(),
-        config.play_pause.key(),
+        data.play_pause.modifiers(),
+        data.play_pause.key(),
         Event::PlayPause,
     );
-    hk.register_hotkey(config.quit.modifiers(), config.quit.key(), Event::Quit);
+    hk.register_hotkey(data.quit.modifiers(), data.quit.key(), Event::Quit);
 
     loop {
         if let Some(event) = hk.listen() {
